LISA class (Internally Used)

A LISA class wrappers the results of LISA computation

Fields

gda_lisa

An object of GeoDaLISA

p_vals

The pseudo-p values of significance of LISA computation

c_vals

The cluster indicators of LISA computation

lisa_vals

The local spatial autocorrelation values of LISA computation

nn_vals

The number of neighbors of every observations in LISA computation

lag_vals

The values of spatial lags of LISA computation

num_obs

Number of observations

labels

The cluster labels of LISA

colors

The cluster colors (HEX format) of LISA

Methods

GetBO(current_p)

Get the Bonferroni bound value

GetClusterIndicators()

Get the local cluster indicators returned from LISA computation.

GetColors()

Get the cluster colors of LISA computation.

GetFDR(current_p)

Get the False Discovery Rate value

GetLISAValues()

Get the local spatial autocorrelation values returned from LISA computation.

GetLabels()

Get the cluster labels of LISA computation.

GetLocalSignificanceValues()

Get the local pseudo-p values of significance returned from LISA computation.

GetNumNeighbors()

Get the number of neighbors of every observations in LISA computation.

Run()

Call to run LISA computation

SetPermutations(num_perm)

Set the number of permutations for the LISA computation

SetSignificanceCutoff(cutoff)

Set the cutoff value of significance values

SetThreads(num_threads)

Set the number of CPU threads for the LISA computation

initialize(lisa_obj)

Constructor with a LISA object (internally used)