Create a geoda object by reading a spatial dataset

Create a geoda object by reading a spatial dataset. The dataset that rgeoda supports includes: ESRI Shapefile, MapInfo File, CSV, GML, GPX, KML, GeoJSON, TopoJSON, OpenFileGDB, GFT Google Fusion Tables, CouchDB

geoda_open(ds_path)

Arguments

ds_path
(character) The path of the spatial dataset

Value

gda_obj An object of geoda instance

Examples

guerry_path <- system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry <- geoda_open(guerry_path)
guerry_df <- as.data.frame(guerry) # access as a data.frame
head(guerry_df)

#>      CODE_DE COUNT AVE_ID_ dept Region      Dprtmnt Crm_prs Crm_prp Litercy
#> NA        01     1      49    1      E          Ain   28870   15890      37
#> NA.1      02     1     812    2      N        Aisne   26226    5521      51
#> NA.2      03     1    1418    3      C       Allier   26747    7925      13
#> NA.3      04     1    1603    4      E Basses-Alpes   12935    7289      46
#> NA.4      05     1    1802    5      E Hautes-Alpes   17488    8174      69
#> NA.5      07     1    2249    7      S      Ardeche    9474   10263      27
#>      Donatns Infants Suicids MainCty Wealth Commerc Clergy Crm_prn Infntcd
#> NA      5098   33120   35039       2     73      58     11      71      60
#> NA.1    8901   14572   12831       2     22      10     82       4      82
#> NA.2   10973   17044  114121       2     61      66     68      46      42
#> NA.3    2733   23018   14238       1     76      49      5      70      12
#> NA.4    6962   23076   16171       1     83      65     10      22      23
#> NA.5    3188   42117   52547       1     84       1     28      76      47
#>      Dntn_cl Lottery Desertn Instrct Prsttts Distanc Area Pop1831 TopCrm TopLit
#> NA        69      41      55      46      13 218.372 5762  346.03      1      0
#> NA.1      36      38      82      24     327  65.945 7369  513.00      1      1
#> NA.2      76      66      16      85      34 161.927 7340  298.26      1      0
#> NA.3      37      80      32      29       2 351.399 6925  155.90      0      1
#> NA.4      64      79      35       7       1 320.280 5549  129.10      0      1
#> NA.5      67      70      19      62       1 279.413 5529  340.73      0      0
#>      TopWealth
#> NA           1
#> NA.1         0
#> NA.2         1
#> NA.3         1
#> NA.4         1
#> NA.5         1