Regionalization with dynamically constrained agglomerative clustering and partitioning

REDCAP (Regionalization with dynamically constrained agglomerative clustering and partitioning) is developed by D. Guo (2008). Like SKATER, REDCAP starts from building a spanning tree with 4 different ways (single-linkage, average-linkage, ward-linkage and the complete-linkage). The single-linkage way leads to build a minimum spanning tree. Then,REDCAP provides 2 different ways (first-order and full-order constraining) to prune the tree to find clusters. The first-order approach with a minimum spanning tree is exactly the same with SKATER. In GeoDa and pygeoda, the following methods are provided: \* First-order and Single-linkage \* Full-order and Complete-linkage \* Full-order and Average-linkage \* Full-order and Single-linkage \* Full-order and Ward-linkage

redcap(
  k,
  w,
  data,
  method = "fullorder-averagelinkage",
  bound_vals = vector("numeric"),
  min_bound = 0,
  distance_method = "euclidean",
  random_seed = 123456789,
  cpu_threads = 6
)

Arguments

k
The number of clusters

w
An instance of Weight class

data
A list of numeric vectors of selected variable

method
"firstorder-singlelinkage", "fullorder-completelinkage", "fullorder-averagelinkage","fullorder-singlelinkage", "fullorder-wardlinkage"

bound_vals
(optional) A 1-d vector of selected bounding variable

min_bound
(optional) A minimum value that the sum value of bounding variable int each cluster should be greater than

distance_method
(optional) The distance method used to compute the distance betwen observation i and j. Defaults to "euclidean". Options are "euclidean" and "manhattan"

random_seed
(int,optional) The seed for random number generator. Defaults to 123456789.

cpu_threads
(optional) The number of cpu threads used for parallel computation

Value

A list of numeric vectors represents a group of clusters

Examples

guerry_path <- system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry <- geoda_open(guerry_path)
queen_w <- queen_weights(guerry)
guerry_df <- as.data.frame(guerry) # use as data.frame
data <- guerry_df[c('Crm_prs','Crm_prp','Litercy','Donatns','Infants','Suicids')]
guerry_clusters <- redcap(4, queen_w, data, "fullorder-completelinkage")
guerry_clusters

#> [[1]]
#>  [1] 46 50 34 38 15 74 21 16 55 13 60 69 47 58 19 32 33 53 39 68 17 82  0 41 81
#> [26]  2 14  5 26 80 20
#> 
#> [[2]]
#>  [1] 49 52 72 84 56 59  1 57 25 42  8 48 54 64 51 18 36 73 24 71  6 67 83 75 65
#> [26] 22 70 12
#> 
#> [[3]]
#> [1] 40
#> 
#> [[4]]
#>  [1] 23 79 29 61 44  3 28 76 77  9 27 45 11 78 43 35 66 31 30 37 10 63  4  7 62
#>