Spatially Constrained Hierarchical Clucstering (SCHC)

Spatially constrained hierarchical clustering is a special form of constrained clustering, where the constraint is based on contiguity (common borders). The method builds up the clusters using agglomerative hierarchical clustering methods: single linkage, complete linkage, average linkage and Ward's method (a special form of centroid linkage). Meanwhile, it also maintains the spatial contiguity when merging two clusters.

schc(
  k,
  w,
  data,
  method = "average",
  bound_vals = vector("numeric"),
  min_bound = 0,
  distance_method = "euclidean"
)

Arguments

k
The number of clusters

w
An instance of Weight class

data
A list of numeric vectors of selected variable

method
"single", "complete", "average","ward"

bound_vals
(optional) A 1-d vector of selected bounding variable

min_bound
(optional) A minimum value that the sum value of bounding variable int each cluster should be greater than

distance_method
(optional) The distance method used to compute the distance betwen observation i and j. Defaults to "euclidean". Options are "euclidean" and "manhattan"

Value

A list of numeric vectors represents a group of clusters

Examples

guerry_path <- system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry <- geoda_open(guerry_path)
queen_w <- queen_weights(guerry)
guerry_df <- as.data.frame(guerry) # use as data.frame
data <- guerry_df[c('Crm_prs','Crm_prp','Litercy','Donatns','Infants','Suicids')]
guerry_clusters <- schc(4, queen_w, data, "complete")
guerry_clusters

#> [[1]]
#> [1] 40
#> 
#> [[2]]
#> [1] 20
#> 
#> [[3]]
#> [1] 80
#> 
#> [[4]]
#>  [1]  0  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 35 36 37 38 39 41 42 43 44 45 46 47 48 49 50 51
#> [51] 52 53 54 55 56 57 58 59 60 61 62 63 64 65 66 67 68 69 70 71 72 73 74 75 76
#> [76] 77 78 79 81 82 83 84
#>